change2

+++ Tamamushi/TMGradientNavigationBar.swift
@@ -66,7 +66,7 @@ public class TMGradientNavigationBar: NSObject {
     }
 
     // Create UIImage with the gradient specified
-    public func generateGradientImage(direction: Direction, startColor: UIColor, endColor: UIColor, startPoint: CGPoint, endPoint: CGPoint, height: Int = 64) -> UIImage {
+    public func generateGradientImage(direction: Direction, startColor: UIColor, endColor: UIColor, startPoint: CGPoint, endPoint: CGPoint, height: CGFloat = 64.0) -> UIImage {
         let gradientLayer = CAGradientLayer()
         //    gradient.frame = CGRect(x: 0, y: 0, width: UIApplication.sharedApplication().statusBarFrame.width, height: UIApplication.sharedApplication().statusBarFrame.height + self.navigationController!.navigationBar.frame.height)
